==> 03_E1_tune_model_lenskit.ipynb <==
cd <your path>

# Begining

## Load modules

! pip install pyspark

from operator import itemgetter
import time
import sys
import getpass
from pathlib import Path
import pandas as pd
import numpy as np

import itertools

## Define the Parameters

# PATH = '/scratch/work/courses/DSGA1004-2021/MSD'
PATH = './data'

STATS_PATH = Path('./saved_tuned_model_stats')

WHOLE_DATA = True
FRACTION = 1.00
TOP=500
PREC_AT = 500

# Load the data

if WHOLE_DATA:
    train_pd = pd.read_parquet(PATH+'/cf_train_trans.parquet')
    unique_user_index_val_pd = pd.read_parquet(PATH+'/unique_user_index_val.parquet')
else:
    train_pd = pd.read_parquet(PATH+'/cf_train_trans_{}.parquet'.format(str(FRACTION).replace('.','_')))
    unique_user_index_val_pd = pd.read_parquet(PATH+'/unique_user_index_val.parquet')
    unique_user_index_val_pd = unique_user_index_val_pd.sample(replace=False, frac=max(FRACTION,0.05)) 

val_pd = pd.read_parquet(PATH+'/cf_validation_trans.parquet')
print('Successfully loaded the data')

train_pd.rename(columns={'user_index':'user','track_index':'item','count':'rating'}, inplace=True)
train_pd.head(2)

unique_user_index_val_pd = np.array(unique_user_index_val_pd.values.tolist())
unique_user_index_val_list = list(unique_user_index_val_pd.reshape(-1))

val_pd.rename(columns={'user_index':'user','track_index':'item','count':'rating'}, inplace=True)
val_pd.head(2)

# ALS Model lenskit

! pip install lenskit

from lenskit.algorithms import als as ALS_ext
from lenskit.algorithms import Recommender
from lenskit import batch, topn, util

ranks = [i for i in range(10,40,10)]
# ranks = [i for i in range(100,500,30)]
regParams = [1] #[10**i for i in range(-2,1)   
alphas = [1]
maxIter = [5]

params = [[a,b,c,d] for a,b,c,d in itertools.product(ranks, regParams, alphas, maxIter)]
print('length of params',len(params))

precisionAt_k_highest = 0
stats_file = open(STATS_PATH/'stats_lenskit_{}.txt'.format(str(FRACTION).replace('.','_')), 'a', buffering=1)
print(' '.join(sys.argv))
print(' '.join(sys.argv), file=stats_file)
for rk,reg,alp,it in itertools.product(ranks, regParams, alphas, maxIter):
    since = time.time()
    als_ext = ALS_ext.ImplicitMF(features=rk, iterations=it, reg=reg, weight=alp)
    als_ext = Recommender.adapt(als_ext)

    model_ext = als_ext.fit(train_pd)
    time_to_fit = time.time() - since

    pred_recs_tracks_val_ext = batch.recommend(model_ext, unique_user_index_val_list, TOP)
    time_to_recommend = time.time() - since

    rla = topn.RecListAnalysis()
    rla.add_metric(topn.precision)
    results = rla.compute(pred_recs_tracks_val_ext, val_pd)
    precisionAt_k = results.precision.mean()
    time_to_eval = time.time() - since
    
    print("With Rank:{}, Reg:{}, Alpha:{}, Maxiter:{}, Metric: {}, time TTL: {}, fit: {}, rec: {}, evl: {}".format(
        rk,reg,alp,it, 
        precisionAt_k, 
        time_to_eval, time_to_fit, time_to_recommend-time_to_fit, time_to_eval-time_to_recommend 
    ), file=stats_file)

    print("With Rank:{}, Reg:{}, Alpha:{}, Maxiter:{}, Metric: {}, time TTL: {}, fit: {}, rec: {}, evl: {}".format(
        rk,reg,alp,it, 
        precisionAt_k, 
        time_to_eval, time_to_fit, time_to_recommend-time_to_fit, time_to_eval-time_to_recommend 
    ))

ranks = [i for i in range(100,320,30)]
# ranks = [i for i in range(100,500,30)]
regParams = [1] #[10**i for i in range(-2,1)   
alphas = [1]
maxIter = [5]

params = [[a,b,c,d] for a,b,c,d in itertools.product(ranks, regParams, alphas, maxIter)]
print('length of params',len(params))

precisionAt_k_highest = 0
stats_file = open(STATS_PATH/'stats_lenskit_{}.txt'.format(str(FRACTION).replace('.','_')), 'a', buffering=1)
print(' '.join(sys.argv))
print(' '.join(sys.argv), file=stats_file)
for rk,reg,alp,it in itertools.product(ranks, regParams, alphas, maxIter):
    since = time.time()
    als_ext = ALS_ext.ImplicitMF(features=rk, iterations=it, reg=reg, weight=alp)
    als_ext = Recommender.adapt(als_ext)

    model_ext = als_ext.fit(train_pd)
    time_to_fit = time.time() - since

    pred_recs_tracks_val_ext = batch.recommend(model_ext, unique_user_index_val_list, TOP)
    time_to_recommend = time.time() - since

    rla = topn.RecListAnalysis()
    rla.add_metric(topn.precision)
    results = rla.compute(pred_recs_tracks_val_ext, val_pd)
    precisionAt_k = results.precision.mean()
    time_to_eval = time.time() - since
    
    print("With Rank:{}, Reg:{}, Alpha:{}, Maxiter:{}, Metric: {}, time TTL: {}, fit: {}, rec: {}, evl: {}".format(
        rk,reg,alp,it, 
        precisionAt_k, 
        time_to_eval, time_to_fit, time_to_recommend-time_to_fit, time_to_eval-time_to_recommend 
    ), file=stats_file)

    print("With Rank:{}, Reg:{}, Alpha:{}, Maxiter:{}, Metric: {}, time TTL: {}, fit: {}, rec: {}, evl: {}".format(
        rk,reg,alp,it, 
        precisionAt_k, 
        time_to_eval, time_to_fit, time_to_recommend-time_to_fit, time_to_eval-time_to_recommend 
    ))
